Actually made an annoying page Still under construction a lot more to come hopefully.
Relaxing smooth jazz plays with a tin-e audio effect and a button that moves around the screen when the user goes to click on it to leave the page. A counter is incremented everytime it moves and when it gets to 3 it changes the text and stops moving.

Still under construction a lot more to come hopefully.