Adding My Page
What's annoying about my page is that you're trapped in a maze puzzle based off of the gen 1 pokemon's psychic gym.
The game is buggy but there aren't any game breaking bugs I've discovered when testing so reaching the end that will send you back to the directory will be hard but not impossible.